merger inntektstrategiene da de er avhengig av hverandre
