+++ ytelse-aktivitetspenger/src/main/java/no/nav/ung/ytelse/aktivitetspenger/formidling/vedtak/EndringInntektStrategy.java
@@ -5,9 +5,11 @@ import jakarta.inject.Inject;
 import no.nav.fpsak.tidsserie.LocalDateTimeline;
 import no.nav.fpsak.tidsserie.StandardCombinators;
 import no.nav.ung.kodeverk.behandling.FagsakYtelseType;
+import no.nav.ung.kodeverk.behandling.aksjonspunkt.AksjonspunktDefinisjon;
 import no.nav.ung.kodeverk.dokument.DokumentMalType;
 import no.nav.ung.sak.behandlingskontroll.FagsakYtelseTypeRef;
 import no.nav.ung.sak.behandlingslager.behandling.Behandling;
+import no.nav.ung.sak.behandlingslager.behandling.aksjonspunkt.Aksjonspunkt;
 import no.nav.ung.sak.behandlingslager.tilkjentytelse.KontrollertInntektPeriode;
 import no.nav.ung.sak.behandlingslager.tilkjentytelse.TilkjentYtelseRepository;
 import no.nav.ung.sak.formidling.vedtak.regler.IngenBrevÅrsakType;
@@ -17,20 +19,32 @@ import no.nav.ung.sak.formidling.vedtak.regler.strategy.VedtaksbrevStrategyResul
 import no.nav.ung.sak.formidling.vedtak.resultat.DetaljertResultat;
 import no.nav.ung.sak.formidling.vedtak.resultat.DetaljertResultatType;
 import no.nav.ung.sak.formidling.vedtak.resultat.ResultatHelper;
+import no.nav.ung.ytelse.aktivitetspenger.formidling.innhold.EndringInntektReduksjonInnholdBygger;
 import no.nav.ung.ytelse.aktivitetspenger.formidling.innhold.EndringInntektUtenReduksjonInnholdBygger;
 
 import java.math.BigDecimal;
 import java.util.List;
 
+/**
+ * Samler brevutledning for utfallet av kontroll av inntekt. Reduksjon (eller ingen utbetaling) og full
+ * utbetaling er gjensidig utelukkende utfall av samme flyt, og avgjøres derfor her i én strategi slik at
+ * det aldri produseres både et reduksjonsbrev og et uten-reduksjon-brev for samme behandling. Reduksjon
+ * har presedens over full utbetaling.
+ */
 @Dependent
 @FagsakYtelseTypeRef(FagsakYtelseType.AKTIVITETSPENGER)
-public final class EndringInntektUtenReduksjonStrategy implements VedtaksbrevInnholdbyggerStrategy {
+public final class EndringInntektStrategy implements VedtaksbrevInnholdbyggerStrategy {
 
+    private final EndringInntektReduksjonInnholdBygger endringInntektReduksjonInnholdBygger;
     private final EndringInntektUtenReduksjonInnholdBygger endringInntektUtenReduksjonInnholdBygger;
     private final TilkjentYtelseRepository tilkjentYtelseRepository;
 
     @Inject
-    public EndringInntektUtenReduksjonStrategy(EndringInntektUtenReduksjonInnholdBygger endringInntektUtenReduksjonInnholdBygger, TilkjentYtelseRepository tilkjentYtelseRepository) {
+    public EndringInntektStrategy(
+        EndringInntektReduksjonInnholdBygger endringInntektReduksjonInnholdBygger,
+        EndringInntektUtenReduksjonInnholdBygger endringInntektUtenReduksjonInnholdBygger,
+        TilkjentYtelseRepository tilkjentYtelseRepository) {
+        this.endringInntektReduksjonInnholdBygger = endringInntektReduksjonInnholdBygger;
         this.endringInntektUtenReduksjonInnholdBygger = endringInntektUtenReduksjonInnholdBygger;
         this.tilkjentYtelseRepository = tilkjentYtelseRepository;
     }
@@ -38,16 +52,56 @@ public final class EndringInntektUtenReduksjonStrategy implements VedtaksbrevInn
     @Override
     public List<VedtaksbrevStrategyResultat> evaluer(Behandling behandling, LocalDateTimeline<DetaljertResultat> detaljertResultat) {
         var resultater = new ResultatHelper(VedtaksbrevInnholdbyggerStrategy.tilResultatInfo(detaljertResultat));
-        if (!resultater.innholder(DetaljertResultatType.KONTROLLER_INNTEKT_FULL_UTBETALING)) {
-            return List.of();
+        boolean harReduksjon = resultater.innholder(DetaljertResultatType.KONTROLLER_INNTEKT_REDUKSJON)
+            || resultater.innholder(DetaljertResultatType.KONTROLLER_INNTEKT_INGEN_UTBETALING);
+        boolean harFullUtbetaling = resultater.innholder(DetaljertResultatType.KONTROLLER_INNTEKT_FULL_UTBETALING);
+
+        if (harReduksjon) {
+            return List.of(reduksjonResultat(behandling));
+        }
+        if (harFullUtbetaling) {
+            return List.of(fullUtbetalingResultat(behandling, detaljertResultat));
+        }
+        return List.of();
+    }
+
+    private VedtaksbrevStrategyResultat reduksjonResultat(Behandling behandling) {
+        boolean harUtførtKontrollerInntekt = behandling.getAksjonspunkter().stream()
+            .filter(Aksjonspunkt::erUtført)
+            .anyMatch(it -> it.getAksjonspunktDefinisjon() == AksjonspunktDefinisjon.KONTROLLER_INNTEKT);
+
+        var forklaring = "Automatisk brev ved endring av inntekt.";
+        if (harUtførtKontrollerInntekt) {
+            return medRedigerbarKontrollerInntektBrev(forklaring);
         }
 
+        return VedtaksbrevStrategyResultat.medUredigerbarBrev(DokumentMalType.ENDRING_INNTEKT,
+            endringInntektReduksjonInnholdBygger,
+            forklaring);
+    }
+
+    private VedtaksbrevStrategyResultat medRedigerbarKontrollerInntektBrev(String forklaring) {
+        forklaring += " Kan redigere pga ap=" + AksjonspunktDefinisjon.KONTROLLER_INNTEKT.getKode() + ".";
+        return new VedtaksbrevStrategyResultat(
+            DokumentMalType.ENDRING_INNTEKT,
+            endringInntektReduksjonInnholdBygger,
+            new VedtaksbrevEgenskaper(
+                false,
+                false,
+                true,
+                true),
+            null,
+            forklaring
+        );
+    }
+
+    private VedtaksbrevStrategyResultat fullUtbetalingResultat(Behandling behandling, LocalDateTimeline<DetaljertResultat> detaljertResultat) {
         var kontrollertInntektPerioderTidslinje = hentKontrollertInntektTidslinje(behandling);
 
         var harManueltFastsattInntekt = harManueltFastsattInntekt(detaljertResultat, kontrollertInntektPerioderTidslinje);
 
         if (harManueltFastsattInntekt) {
-            return List.of(new VedtaksbrevStrategyResultat(
+            return new VedtaksbrevStrategyResultat(
                 DokumentMalType.ENDRING_INNTEKT_UTEN_REDUKSJON,
                 endringInntektUtenReduksjonInnholdBygger,
                 new VedtaksbrevEgenskaper(false,
@@ -56,9 +110,9 @@ public final class EndringInntektUtenReduksjonStrategy implements VedtaksbrevInn
                     true),
                 null,
                 "Redigerbar brev ved full utbetaling med manuelt fastsatt inntekt på 0 kr uten registerinntekt."
-            ));
+            );
         }
-        return List.of(VedtaksbrevStrategyResultat.utenBrev(IngenBrevÅrsakType.IKKE_RELEVANT, "Ingen brev ved full utbetaling etter kontroll av inntekt."));
+        return VedtaksbrevStrategyResultat.utenBrev(IngenBrevÅrsakType.IKKE_RELEVANT, "Ingen brev ved full utbetaling etter kontroll av inntekt.");
     }
 
     private boolean harManueltFastsattInntekt(
@@ -91,4 +145,3 @@ public final class EndringInntektUtenReduksjonStrategy implements VedtaksbrevInn
     }
 
 }
-
