--- ytelse-ungdomsprogramytelsen/src/main/java/no/nav/ung/ytelse/ungdomsprogramytelsen/formidling/vedtak/regler/strategy/EndringInntektReduksjonStrategy.java
@@ -1,72 +1,0 @@
-package no.nav.ung.ytelse.ungdomsprogramytelsen.formidling.vedtak.regler.strategy;
-
-import jakarta.enterprise.context.Dependent;
-import jakarta.inject.Inject;
-import no.nav.fpsak.tidsserie.LocalDateTimeline;
-import no.nav.ung.kodeverk.behandling.FagsakYtelseType;
-import no.nav.ung.kodeverk.behandling.aksjonspunkt.AksjonspunktDefinisjon;
-import no.nav.ung.kodeverk.dokument.DokumentMalType;
-import no.nav.ung.sak.behandlingskontroll.FagsakYtelseTypeRef;
-import no.nav.ung.sak.behandlingslager.behandling.Behandling;
-import no.nav.ung.sak.behandlingslager.behandling.aksjonspunkt.Aksjonspunkt;
-import no.nav.ung.sak.formidling.vedtak.regler.VedtaksbrevEgenskaper;
-import no.nav.ung.sak.formidling.vedtak.regler.strategy.VedtaksbrevInnholdbyggerStrategy;
-import no.nav.ung.sak.formidling.vedtak.regler.strategy.VedtaksbrevStrategyResultat;
-import no.nav.ung.sak.formidling.vedtak.resultat.DetaljertResultat;
-import no.nav.ung.sak.formidling.vedtak.resultat.DetaljertResultatType;
-import no.nav.ung.sak.formidling.vedtak.resultat.ResultatHelper;
-import no.nav.ung.ytelse.ungdomsprogramytelsen.formidling.innhold.EndringInntektReduksjonInnholdBygger;
-
-import java.util.List;
-
-@Dependent
-@FagsakYtelseTypeRef(FagsakYtelseType.UNGDOMSYTELSE)
-public final class EndringInntektReduksjonStrategy implements VedtaksbrevInnholdbyggerStrategy {
-
-    private final EndringInntektReduksjonInnholdBygger endringInntektReduksjonInnholdBygger;
-
-    @Inject
-    public EndringInntektReduksjonStrategy(EndringInntektReduksjonInnholdBygger endringInntektReduksjonInnholdBygger) {
-        this.endringInntektReduksjonInnholdBygger = endringInntektReduksjonInnholdBygger;
-    }
-
-    @Override
-    public List<VedtaksbrevStrategyResultat> evaluer(Behandling behandling, LocalDateTimeline<DetaljertResultat> detaljertResultat) {
-        var resultater = new ResultatHelper(VedtaksbrevInnholdbyggerStrategy.tilResultatInfo(detaljertResultat));
-        boolean erRelevant = resultater.innholder(DetaljertResultatType.KONTROLLER_INNTEKT_REDUKSJON)
-            || resultater.innholder(DetaljertResultatType.KONTROLLER_INNTEKT_INGEN_UTBETALING);
-        if (!erRelevant) {
-            return List.of();
-        }
-
-        boolean harUtførtKontrollerInntekt = behandling.getAksjonspunkter().stream()
-            .filter(Aksjonspunkt::erUtført)
-            .anyMatch(it -> it.getAksjonspunktDefinisjon() == AksjonspunktDefinisjon.KONTROLLER_INNTEKT);
-
-        var forklaring = "Automatisk brev ved endring av inntekt.";
-        if (harUtførtKontrollerInntekt) {
-            return List.of(medRedigerbarKontrollerInntektBrev(forklaring));
-        }
-
-        return List.of(VedtaksbrevStrategyResultat.medUredigerbarBrev(DokumentMalType.ENDRING_INNTEKT,
-            endringInntektReduksjonInnholdBygger,
-            forklaring));
-    }
-
-    private VedtaksbrevStrategyResultat medRedigerbarKontrollerInntektBrev(String forklaring) {
-        forklaring += " Kan redigere pga ap=" + AksjonspunktDefinisjon.KONTROLLER_INNTEKT.getKode() + ".";
-        return new VedtaksbrevStrategyResultat(
-            DokumentMalType.ENDRING_INNTEKT,
-            endringInntektReduksjonInnholdBygger,
-            new VedtaksbrevEgenskaper(
-                false,
-                false,
-                true,
-                true),
-            null,
-            forklaring
-        );
-    }
-
-}
-
