+++ ytelse-ungdomsprogramytelsen/src/main/java/no/nav/ung/ytelse/ungdomsprogramytelsen/formidling/vedtak/regler/strategy/EndringInntektStrategy.java
@@ -20,6 +20,7 @@ import no.nav.ung.sak.formidling.vedtak.regler.strategy.VedtaksbrevStrategyResul
 import no.nav.ung.sak.formidling.vedtak.resultat.DetaljertResultat;
 import no.nav.ung.sak.formidling.vedtak.resultat.DetaljertResultatType;
 import no.nav.ung.sak.formidling.vedtak.resultat.ResultatHelper;
+import no.nav.ung.ytelse.ungdomsprogramytelsen.formidling.innhold.EndringInntektReduksjonInnholdBygger;
 import no.nav.ung.ytelse.ungdomsprogramytelsen.formidling.innhold.EndringInntektUtenReduksjonInnholdBygger;
 
 import java.math.BigDecimal;
@@ -27,16 +28,21 @@ import java.util.List;
 
 @Dependent
 @FagsakYtelseTypeRef(FagsakYtelseType.UNGDOMSYTELSE)
-public final class EndringInntektUtenReduksjonStrategy implements VedtaksbrevInnholdbyggerStrategy {
-
+public final class EndringInntektStrategy implements VedtaksbrevInnholdbyggerStrategy {
 
     private final boolean enableEndringUtenReduksjonSjekk;
+    private final EndringInntektReduksjonInnholdBygger endringInntektReduksjonInnholdBygger;
     private final EndringInntektUtenReduksjonInnholdBygger endringInntektUtenReduksjonInnholdBygger;
     private final TilkjentYtelseRepository tilkjentYtelseRepository;
 
     @Inject
-    public EndringInntektUtenReduksjonStrategy(@KonfigVerdi(value = "ENABLE_ENDRING_UTEN_REDUKSJON_SJEKK", defaultVerdi = "false") boolean enableEndringUtenReduksjonSjekk, EndringInntektUtenReduksjonInnholdBygger endringInntektUtenReduksjonInnholdBygger, TilkjentYtelseRepository tilkjentYtelseRepository) {
+    public EndringInntektStrategy(
+        @KonfigVerdi(value = "ENABLE_ENDRING_UTEN_REDUKSJON_SJEKK", defaultVerdi = "false") boolean enableEndringUtenReduksjonSjekk,
+        EndringInntektReduksjonInnholdBygger endringInntektReduksjonInnholdBygger,
+        EndringInntektUtenReduksjonInnholdBygger endringInntektUtenReduksjonInnholdBygger,
+        TilkjentYtelseRepository tilkjentYtelseRepository) {
         this.enableEndringUtenReduksjonSjekk = enableEndringUtenReduksjonSjekk;
+        this.endringInntektReduksjonInnholdBygger = endringInntektReduksjonInnholdBygger;
         this.endringInntektUtenReduksjonInnholdBygger = endringInntektUtenReduksjonInnholdBygger;
         this.tilkjentYtelseRepository = tilkjentYtelseRepository;
     }
@@ -44,16 +50,57 @@ public final class EndringInntektUtenReduksjonStrategy implements VedtaksbrevInn
     @Override
     public List<VedtaksbrevStrategyResultat> evaluer(Behandling behandling, LocalDateTimeline<DetaljertResultat> detaljertResultat) {
         var resultater = new ResultatHelper(VedtaksbrevInnholdbyggerStrategy.tilResultatInfo(detaljertResultat));
-        if (!resultater.innholder(DetaljertResultatType.KONTROLLER_INNTEKT_FULL_UTBETALING)) {
-            return List.of();
+        boolean harReduksjon = resultater.innholder(DetaljertResultatType.KONTROLLER_INNTEKT_REDUKSJON)
+            || resultater.innholder(DetaljertResultatType.KONTROLLER_INNTEKT_INGEN_UTBETALING);
+        boolean harFullUtbetaling = resultater.innholder(DetaljertResultatType.KONTROLLER_INNTEKT_FULL_UTBETALING);
+
+        if (harReduksjon) {
+            return List.of(reduksjonResultat(behandling));
+        }
+        // Kun hvis har ikke reduksjon
+        if (harFullUtbetaling) {
+            return List.of(fullUtbetalingResultat(behandling, detaljertResultat));
+        }
+        return List.of();
+    }
+
+    private VedtaksbrevStrategyResultat reduksjonResultat(Behandling behandling) {
+        boolean harUtførtKontrollerInntekt = behandling.getAksjonspunkter().stream()
+            .filter(Aksjonspunkt::erUtført)
+            .anyMatch(it -> it.getAksjonspunktDefinisjon() == AksjonspunktDefinisjon.KONTROLLER_INNTEKT);
+
+        var forklaring = "Automatisk brev ved endring av inntekt.";
+        if (harUtførtKontrollerInntekt) {
+            return medRedigerbarKontrollerInntektBrev(forklaring);
         }
 
+        return VedtaksbrevStrategyResultat.medUredigerbarBrev(DokumentMalType.ENDRING_INNTEKT,
+            endringInntektReduksjonInnholdBygger,
+            forklaring);
+    }
+
+    private VedtaksbrevStrategyResultat medRedigerbarKontrollerInntektBrev(String forklaring) {
+        forklaring += " Kan redigere pga ap=" + AksjonspunktDefinisjon.KONTROLLER_INNTEKT.getKode() + ".";
+        return new VedtaksbrevStrategyResultat(
+            DokumentMalType.ENDRING_INNTEKT,
+            endringInntektReduksjonInnholdBygger,
+            new VedtaksbrevEgenskaper(
+                false,
+                false,
+                true,
+                true),
+            null,
+            forklaring
+        );
+    }
+
+    private VedtaksbrevStrategyResultat fullUtbetalingResultat(Behandling behandling, LocalDateTimeline<DetaljertResultat> detaljertResultat) {
         var kontrollertInntektPerioderTidslinje = hentKontrollertInntektTidslinje(behandling);
 
         var harManueltFastsattInntekt = harManueltFastsattInntekt(behandling, detaljertResultat, kontrollertInntektPerioderTidslinje);
 
         if (harManueltFastsattInntekt) {
-            return List.of(new VedtaksbrevStrategyResultat(
+            return new VedtaksbrevStrategyResultat(
                 DokumentMalType.ENDRING_INNTEKT_UTEN_REDUKSJON,
                 endringInntektUtenReduksjonInnholdBygger,
                 new VedtaksbrevEgenskaper(false,
@@ -62,9 +109,9 @@ public final class EndringInntektUtenReduksjonStrategy implements VedtaksbrevInn
                     true),
                 null,
                 "Redigerbar brev ved full utbetaling med manuelt fastsatt inntekt på 0 kr uten registerinntekt."
-            ));
+            );
         }
-        return List.of(VedtaksbrevStrategyResultat.utenBrev(IngenBrevÅrsakType.IKKE_RELEVANT, "Ingen brev ved full utbetaling etter kontroll av inntekt."));
+        return VedtaksbrevStrategyResultat.utenBrev(IngenBrevÅrsakType.IKKE_RELEVANT, "Ingen brev ved full utbetaling etter kontroll av inntekt.");
     }
 
     private boolean harManueltFastsattInntekt(
